+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @inlang/core
 
+## 0.7.9
+
+### Patch Changes
+
+- f0cc15bb: add missing `PluginSetupFunction` export
+
+## 0.7.1
+
+### Patch Changes
+
+- ab2d90cf: fix `pluginBuildConfig`
+
+## 0.7.0
+
+### Minor Changes
+
+- a5572e12: Introduction of the plugin API in @inlang/core.
+
 ## 0.6.0
 
 ### Minor Changes