+++ packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/sdk-js
 
+## 0.0.4
+
+### Patch Changes
+
+- dda53294: bundle `sdkPlugin` file
+
+## 0.0.3
+
+### Patch Changes
+
+- f0cc15bb: expose `sdkPlugin` for easier setup
+
 ## 0.0.2
 
 ### Patch Changes